+++ NEWS.md
@@ -4,6 +4,7 @@
 * New argument addcode in make_forest_plot() which inserts code into the generated plot code.
 * New arguments addaes and addarg in make_shape_plot() and make_forest_plot() which can be used to specify additional aesthetics and arguments for some ggplot layers.
 * New argument col.right.parse in make_forest_plot() which controls if columns are parsed into expressions when plotted.
+* New argument `row.labels.levels` in make_forest_plot() and make_forest_data() to specify which columns of row.labels data frame to use. And better handling of missing values.
 * make_forest_plot() no longer returns data. But the plot data is available at .$plot$data
 * New argument envir added to make_shape_plot (allows the user to specify the environment for evaluating the plot code).
 * ggtext::element_markdown() now used for y-axis labels in make_forest_plot()